services/ha-bridge/requirements.txt
fastapi
uvicorn[standard]
aiohttp
paho-mqtt>=2.0
loguru
python-dotenv
